toolchain: don't use -fno-builtin.
This flag, since it removes builtin function usage entirely, can remove
opportunity for potential compiler optimizations to be applied in some
places, and, more importantly, some extra diagnostic warnings that can
appear when the compiler uses the builtin versions of functions (which
is what happened to point out the issues in [1]).

Unfortunately, removing the flag breaks our build. Using
-fno-builtin-printf is enough to fix it, so a safe assumption is that
when using builtin printf(), the compiler can switch it for different
functions (say, puts()), which we don't have our own implementation for,
meaning the libc version is used and that one requires some extra
functions for the underlying stdio implementation to work.

[1] https://github.com/lnls-dig/openMMC/issues/125